Hitboxes and Pseudo Code
I added in functionality and a basic pipeline for hitboxes. Check when an animation gets to a specific frame, then spawn a hitbox with some specific variables, damage, player, size, position etc.
I also made some basic pseudocode for a match class that holds info about the players like the characters and their health and the match status like the timer and the rounds won.